+++ app/src/main/java/com/example/handappmobile_epn/MainActivity.kt
@@ -68,6 +68,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.handappmobile_epn.bt.BluetoothConnectionManager
 import com.example.handappmobile_epn.bt.BluetoothHelper
+import com.example.handappmobile_epn.ui.components.DrawerScreen
 import com.example.handappmobile_epn.ui.components.ToolBar
 import com.example.handappmobile_epn.ui.screen.HomeScreen
 import com.example.handappmobile_epn.ui.screen.PantallaTutorial
@@ -107,94 +108,6 @@ fun ViewContainer(bluetoothConnectionManager: BluetoothConnectionManager) {
         ) {
             DrawerScreen(bluetoothConnectionManager)
         }
-    }
-}
-
-@Composable
-fun DrawerScreen(bluetoothConnectionManager: BluetoothConnectionManager) {
-    var drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
-
-    ModalNavigationDrawer(
-        drawerState = drawerState,
-        drawerContent = { DrawerContent() },
-        content = { ScaffoldContent(bluetoothConnectionManager = bluetoothConnectionManager, drawerState) }
-    )
-}
-
-@Preview
-@Composable
-fun DrawerContent() {
-    val logo = painterResource(id = R.drawable.logocircular)
-
-    Column(modifier = Modifier
-        .background(colorResource(id = R.color.white))
-        .fillMaxHeight()
-        .width(300.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(colorResource(id = R.color.app_dark))
-                .padding(16.dp)
-                .padding(
-                    top = WindowInsets.systemBars
-                        .asPaddingValues()
-                        .calculateTopPadding()
-                ),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Image(
-                painter = logo,
-                contentDescription = "App Logo",
-                modifier = Modifier.size(40.dp)
-            )
-            Spacer(modifier = Modifier.width(16.dp))
-            Text(
-                text = "Hand App",
-                style = MaterialTheme.typography.titleLarge,
-                color = Color.White
-            )
-        }
-
-        MenuItem(text = "Inicio", icon = Icons.Filled.Home, onClick = { /*TODO*/ })
-        MenuItem(text = "Dispositivos", icon = Icons.Filled.Bluetooth, onClick = { /*TODO*/ })
-        MenuItem(text = "Tutorial", icon = Icons.Filled.HelpOutline, onClick = { /*TODO*/ })
-
-        HorizontalDivider()
-
-        MenuItem(text = "Ajustes", icon = Icons.Filled.Settings, onClick = { /*TODO*/ })
-        MenuItem(text = "Acerca de", icon = Icons.Filled.Info, onClick = { /*TODO*/ })
-    }
-}
-
-@Composable
-fun MenuItem(
-    text: String,
-    icon: ImageVector,
-    onClick: () -> Unit
-) {
-    TextButton(onClick = onClick) {
-        Icon(icon, contentDescription = text)
-        Spacer(modifier = Modifier.width(32.dp))
-        Text(text = text)
-    }
-}
-
-@Composable
-fun ScaffoldContent(bluetoothConnectionManager: BluetoothConnectionManager, drawerState: DrawerState) {
-    Scaffold(modifier = Modifier.fillMaxSize(),
-        topBar = { ToolBar(drawerState) }
-    ) { innerPadding ->
-
-        Column(
-            modifier = Modifier.verticalScroll(rememberScrollState())
-        ) {
-            HomeScreen(
-                bluetoothConnectionManager,
-                modifier = Modifier.padding(innerPadding)
-            )
-        }
-
     }
 }
